Separate saving and loading code to saveLoad.ts
