+++ webapp/scripts/saveLoad.ts
@@ -1,0 +1,102 @@
+import {displayError} from "./interface";
+import {
+    getCurrentLanguage,
+    lastNodeString,
+    loadTreeFromString,
+    setCurrentLanguage,
+    setSelectedMode
+} from "./treeManipulation";
+import {getSelectedMode} from "./utils";
+
+const fileInput: HTMLInputElement = document.createElement('input');
+
+export function saveTree(): void {
+    const contents = JSON.stringify({
+        nodeString: lastNodeString,
+        lang: getCurrentLanguage(),
+        mode: getSelectedMode(),
+    })
+    const blob = new Blob([contents], {type: 'text/plain'});
+    const url = window.URL.createObjectURL(blob);
+
+    const a = document.createElement('a');
+    a.href = url;
+    a.download = 'tree.cdtree';
+    document.body.appendChild(a);
+    a.click();
+    document.body.removeChild(a);
+
+    window.URL.revokeObjectURL(url);
+}
+
+export function setupFileInput(): void {
+    fileInput.type = 'file';
+    fileInput.accept = '.cdtree';
+    fileInput.onchange = (event) => {
+        const files = (event.target as HTMLInputElement).files;
+        if (!files || files.length === 0) throw new Error('No file selected');
+        const file = files[0];
+        const reader = new FileReader();
+        reader.onload = () => {
+            loadFromFile(reader);
+        };
+        reader.readAsText(file);
+    };
+}
+
+export function setupFileDragAndDrop(): void {
+    const treeContainer = document.getElementById('tree-container');
+    if (!treeContainer) {
+        console.error('Tree container not found');
+        return;
+    }
+
+    const highlightClass: string = 'file-drag-highlight';
+    const addHighlight = () => treeContainer.classList.add(highlightClass);
+    const removeHighlight = () => treeContainer.classList.remove(highlightClass);
+
+    treeContainer.addEventListener('dragover', (event) => {
+        event.preventDefault();
+        addHighlight();
+    });
+
+    treeContainer.addEventListener('dragleave', removeHighlight);
+
+    treeContainer.addEventListener('drop', (event) => {
+        event.preventDefault();
+        removeHighlight();
+
+        const file: File | undefined = event.dataTransfer?.files[0];
+        if (!file) {
+            displayError('No file dropped');
+            return;
+        }
+
+        const reader = new FileReader();
+        reader.onload = () => loadFromFile(reader);
+        reader.readAsText(file);
+    });
+}
+
+function loadFromFile(reader: FileReader): void {
+    reader.onerror = () => displayError(new Error('Error occurred while attempting to read file'))
+    try {
+        const contents: string = reader.result as string;
+        const json = JSON.parse(contents);
+        if (!json.nodeString || !json.lang || !json.mode) {
+            throw new Error('Provided file did not contain required tree data');
+        }
+        setCurrentLanguage(json.lang);
+        setSelectedMode(json.mode);
+        loadTreeFromString(json.nodeString);
+    } catch (e) {
+        if (e instanceof SyntaxError) {
+            e = new SyntaxError('Provided file was not valid JSON');
+        }
+        displayError(e);
+    }
+}
+
+export function loadTree(): void {
+    fileInput.click();
+}
